Added context menu to current song info
Disabled text selection for current song info, so the main window can be
dragged from empty space on the toolbar. Added a context menu to replace
the original functionality, with the ability to copy information about
the song to the clipboard, and to quickly browse by the playing artist
or the playing album.

https://bugzilla.gnome.org/show_bug.cgi?id=678938